Removing useless files and moving components into folders
