+++ src/app/constants/routesPath.ts
@@ -5,7 +5,6 @@ const routesPaths = {
   verticals: '/verticals/:verticalId',
   nftDetails: '/verticals/:verticalId/:nftId',
   nftById: '/nft/:ftxId',
-  my_collection: '/collection',
 
   knowHow: '/#knowHow',
   creaturesCarousel: '/#carousel',
